AI/requirements.txt
streamlit>=1.30.0
ultralytics>=8.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
numpy>=1.24.0
